TASK04: Implement kobject "wimbledon" with sysfs attribute "list"
On read from the “list” attr. it shows content of the linked list.
On write to the “list” attr. it adds new string to the linked list.

Since the list items separator is not specified - content is shown as it
was inserted including line endings. This gives flexibility of using
"echo" with "-n" argument and without.

Signed-off-by: Ivan Stepanenko <istepanenko@gmail.com>