TASK4: Implement sysfs attributes example [1/1]
Implement a kernel module which contains a
kobject instance containing a linked list and
having a sysfs attribute named "list". Writing
to it should create a new node holding the
string received from user and attach it to the
aforementioned list. Reading the attribute
should print all the exisitng nodes in the list.

Signed-off-by: Victor Krasnoshchok <ct3da21164@protonmail.ch>